WCRP Updates November 2022 (#316)
* WCRP Updates

Update to 4.3. Other change suggestions to follow from Board/WCAT discussions. Then changes to be presented to Staff.

* Reduce proximity restriction from 100km to 10km

* Reduce simultaneous FM competition proximity

From 100km to 10km

* Series competitions issues

Change the complete removal of results to keeping the first competition.
Delegate discretion allowed to remove all results - for example if the competitor is suspected of going to both on purpose instead of an accident.

* Clarify competitors have the right to one guest

* Allow FM champs to be held separately

Includes allowing this for Major Championships.

* Allow later addition to a series

* Specify qualifications are at the end of competitions

* Allow WCAT and Seniors to make minimum competitor limits

* Change 48 hour notice to only be for the first registration

* Allow more registration periods and differences in restrictions based on those periods

* Reduce the requirements for event addition to announced competitions

* Handling additional days of announced competitions

* Reformat wording around series

* Decrease proximity from 19 days to 5 days

Per the Staff feedback period.

* Clarify minimums applied to series